dml: reorg cache related structs
